pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "rpg-tracker"
version = "0.1.0"
description = "Backend-first Django RPG habit tracker API"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "Django>=6.0,<6.1",
    "djangorestframework>=3.17,<3.18",
    "drf-spectacular>=0.28,<0.29",
    "psycopg[binary]>=3.3,<3.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0,<10.0",
    "pytest-django>=4.12,<5.0",
    "ruff>=0.15,<0.16",
]

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
where = ["."]
include = ["config", "config.*", "apps", "apps.*"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

[tool.ruff]
line-length = 100
target-version = "py314"
extend-exclude = ["*/migrations/*.py"]

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
